Canopus:usbhs.h-Modify the return data type of SBHS_HostGetPipeByteCount.
USBHS_HostGetPipeByteCount is used for usb host fifo mode(when
Is_uhd_pipe_dma return false for every pipe)
PBYCT is a 11-bit field.

+++ samv7/libraries/libchip/include/usbhs.h
@@ -1450,7 +1450,7 @@ __STATIC_INLINE uint8_t USBHS_HostGetCurrentBank(Usbhs *pUsbhs, uint8_t Pipe)
  * \brief Gets USB host interrupt status
  * \param pUsbhs  USBHS host instance
  */
-__STATIC_INLINE uint8_t USBHS_HostGetPipeByteCount(Usbhs *pUsbhs, uint8_t Pipe)
+__STATIC_INLINE uint32_t USBHS_HostGetPipeByteCount(Usbhs *pUsbhs, uint8_t Pipe)
 {
 	return ((pUsbhs->USBHS_HSTPIPISR[Pipe] & USBHS_HSTPIPISR_PBYCT_Msk) >>
 			 USBHS_HSTPIPISR_PBYCT_Pos);
